# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User registration",
                "operationId": "register",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RegisterRequest"
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User log in",
                "description": "Returns user data with token",
                "operationId": "login",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User info",
                "description": "Returns user data",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User log out",
                "description": "Method revokes the current access token.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Comments list",
                "description": "Comments list",
                "operationId": "commentsList",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create new comment",
                "operationId": "createComment",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CommentsRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{comment_id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Show comment",
                "operationId": "showComments",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "update comment",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CommentRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "delete comment",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Post list",
                "description": "Post list",
                "operationId": "postList",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create new post",
                "operationId": "createPost",
                "requestBody": {
                    "$ref": "#/components/requestBodies/PostRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post_id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Show post",
                "operationId": "showPost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "update post",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/PostRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "delete post",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/{post_id}/comments": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "show post with comments",
                "operationId": "showPostWithComments",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "requestBodies": {
            "LoginRequest": {
                "description": "Login users.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "email",
                                "password"
                            ],
                            "properties": {
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@email.com"
                                },
                                "password": {
                                    "description": "The user password.",
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RegisterRequest": {
                "description": "Register users.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "email",
                                "password",
                                "password_confirmation"
                            ],
                            "properties": {
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "example": "user@email.com"
                                },
                                "password": {
                                    "description": "The user password.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "123456"
                                },
                                "password_confirmation": {
                                    "description": "The user password confirmation.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "CommentRequest": {
                "description": "Create comment",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "body",
                                "post_id"
                            ],
                            "properties": {
                                "body": {
                                    "description": "Comment body",
                                    "type": "string",
                                    "example": "Some body"
                                },
                                "post_id": {
                                    "description": "post id",
                                    "type": "number",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "PostRequest": {
                "description": "Create post",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "title",
                                "body"
                            ],
                            "properties": {
                                "title": {
                                    "description": "Post title",
                                    "type": "string",
                                    "example": "Some title"
                                },
                                "body": {
                                    "description": "Post body",
                                    "type": "string",
                                    "example": "Some body"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication & Registration"
        },
        {
            "name": "Comments",
            "description": "Comments"
        },
        {
            "name": "Posts",
            "description": "Posts"
        }
    ]
}